Add dice roll command

+++ SaulGoodmanBot/Library/Misc/Poll.cs
@@ -45,7 +45,7 @@ public class Poll {
     }
 
     public string FormatPercent(PollOption option) {
-        return string.Format("{0:P1}", option.Votes / GetTotalVotes());
+        return string.Format("{0:P1}", option.Votes / (double)GetTotalVotes());
     }
 
     public DiscordClient Client { get; set; }
